Deplacement test d'egalite entre les monnaies souhaitée et à vendre dans le service Utilisation de la classe Util pour afficher les messages dans le controller ProposerOffre Changement dans la classe Autorun afin de créer des offres avec des chiffres ayant des décimales aléatoires Correction affichage de certains textes

+++ MasterCoin/src/main/java/fr/ufrima/m2pgi/ecom/service/EchangeOffreService.java
@@ -34,6 +34,10 @@ public class EchangeOffreService {
 		}
 
 		public void addOffre(Compte compte, EchangeOffre eo) throws Exception {
+			if(eo.getMonnaieAchat().equals(eo.getMonnaieVendre())) {
+				context.setRollbackOnly();
+				throw new SameMoneyException();
+			}
 			porteMonnaieService.removeToPorteMonnaie(compte, eo.getMonnaieVendre(), eo.getMontantVendre());
 			eo.setCompte(compte);
 			eo.setDateCreation(new Date());
